split arrays and directories
