+++ src/main/kotlin/me/nobaboy/nobaaddons/core/fishing/TrophyFish.kt
@@ -1,7 +1,7 @@
 package me.nobaboy.nobaaddons.core.fishing
 
 import me.nobaboy.nobaaddons.core.ItemRarity
-import me.nobaboy.nobaaddons.repo.RepoObjects.Companion.listFromRepository
+import me.nobaboy.nobaaddons.repo.RepoObjectArray.Companion.listFromRepository
 import me.nobaboy.nobaaddons.utils.TextUtils.formatted
 import net.minecraft.text.Text
 import net.minecraft.util.Formatting
